Resources Package changed, PathItemFilter Test and Resource
