+++ modules/swagger-core/src/test/java/io/swagger/filter/resources/NoOpenAPIFilter.java
@@ -1,4 +1,4 @@
-package io.swagger.filter;
+package io.swagger.filter.resources;
 
 import io.swagger.core.filter.AbstractSpecFilter;
 import io.swagger.model.ApiDescription;
