+++ modules/swagger-core/src/test/java/io/swagger/filter/resources/ReplaceGetOperationsFilter.java
@@ -1,4 +1,4 @@
-package io.swagger.filter;
+package io.swagger.filter.resources;
 
 import io.swagger.core.filter.AbstractSpecFilter;
 import io.swagger.model.ApiDescription;
@@ -11,11 +11,11 @@ import java.util.Optional;
 /**
  * Sample filter to avoid all get operations for the resource
  **/
-public class NoGetOperationsFilter extends AbstractSpecFilter {
+public class ReplaceGetOperationsFilter extends AbstractSpecFilter {
     @Override
     public Optional<Operation> filterOperation(Operation operation, ApiDescription api, Map<String, List<String>> params, Map<String, String> cookies, Map<String, List<String>> headers) {
         if ("get".equals(api.getMethod())) {
-            return Optional.empty();
+            return Optional.of(new Operation().description("Replaced Operation").operationId("New Operation"));
         }
         return Optional.of(operation);
     }
